ToolshedCheckin: Soft enforce that returnee is same as checkoutee
Add workflow that verifies that the user that last checked out the item
is the user that is checking the item in with the ability to add an
explanation as to why the user does not match or there is no checkout.

Additionally warn on checkouts if the item has already been checked out.

Fixes: https://github.com/LetsBuildConstructionCareers/InventoryScanner/issues/7
Fixes: https://github.com/LetsBuildConstructionCareers/InventoryScanner/issues/9