Fix ViewItemsInContainerActivity does not scroll on items correctly
Previously the UI elements did not take up the full width of the screen
so scrolling would only work if the user toched a UI element like a
picture or a piece of text in order to scroll.

Fix that by making the Rows take up the whole width so scrolling is
possible by touching anywhere on the screen.

Fixes: https://github.com/LetsBuildConstructionCareers/InventoryScanner/issues/14